Toolchain: Backport -ftrivial-auto-var-init to our GCC 11.2 toolchain
This change back ports the patch to support -ftrivial-auto-var-init so
that it compiles on top of the 11.2 release. It was originally committed
to target the GCC 12 release cycle. The back porting was just fixing up
patch collisions, no major surgery was needed.

The original patch can be found here:
https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=a25e0b5e6ac8a77a71c229e0a7b744603365b0e9

Support for applying the patch in our toolchain build was then added to
`BuildIt.sh`.